Including a multi-dimensional array problem
The problem is as follows: We have a matrix m x n, we read the size and
then the matrix itself to load it into memory. Afterwards we take input
of two members of the matrix e.g. 0,0 1,5 and it should sum all the
elements from position 0,0 to position 1,5 linearly.